Move MapData into own file
