+++ Assets/Scripts/Map/MapGeneration/MapData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c9e46d0587862084c8abcae831b87e0b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
